% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{psps}
\alias{psps}
\title{psps}
\usage{
psps(est_lab_y, est_lab_yhat, est_unlab_yhat, Sigma, alpha = 0.05)
}
\arguments{
\item{alpha}{Specifies the confidence level as 1 - alpha for confidence intervals.}

\item{est_lab_y:}{a K-dimensional Array of Point estimates using Y in labeled data.}

\item{est_lab_yhat:}{a K-dimensional Array of Point estimates using Yhat in labeled data.}

\item{est_unlab_yhat:}{a K-dimensional Array of Point estimates using Yhat in unlabeled data.}

\item{Sigma:}{a 3K x 3K Variance-covariance matrix for the above three estimators (Note: not the asymptotic variance).}
}
\value{
A summary table presenting point estimates, standard error, confidence intervals (1 - alpha), P-values for ML-assisted inference.
}
\description{
\code{psps} function Task-Agnostic Machine Learning-Assisted Inference.
}
